# route-service/src/main/java/ru/itmo/soa/route/dto/HumanDto.java
package ru.itmo.soa.route.dto;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
public class HumanDto {
    private String birthday;
}