# route-service/src/main/java/ru/itmo/soa/route/dto/error/ErrorResponse.java
package ru.itmo.soa.route.dto.error;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

import java.util.List;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
public class ErrorResponse {
    private int code;
    private String message;
    private List<String> details;
}